==> 01_eda-global-analysis/02_integration-rivanna.ipynb <==
# TBI snRNA-seq

## Import

import os
from pathlib import Path
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import numpy as np
import scanpy as sc

import random
import torch
import scvi

print(sns.__version__)
print(pd.__version__)
print(np.__version__)
print(sc.__version__)
print(scvi.__version__)

random.seed(0)
torch.manual_seed(0)
np.random.seed(0)
scvi.settings.seed = 0

#os.chdir('../')
#os.getcwd()

data_dir = "2024-06_tbi-snseq-rivanna/h5ad"

adata = sc.read_h5ad(os.path.join(data_dir, '1-tbi-seq-hvg.h5ad'))

adata

## Integration with scVI

### model_A

scvi.model.SCVI.setup_anndata(
    adata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal'],
)

model_A = scvi.model.SCVI(adata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_A.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_A_dir = os.path.join(scvi_dir, 'model_A')
print(model_A_dir)
model_A.save(model_A_dir)

### model_B

scvi.model.SCVI.setup_anndata(
    adata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal', 'doublet_scores'],
)

model_B = scvi.model.SCVI(adata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_B.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_B_dir = os.path.join(scvi_dir, 'model_B')
print(model_B_dir)
model_B.save(model_B_dir)

### model_C

cdata = adata[adata.obs.doublet_scores < 0.3].copy()
cdata

scvi.model.SCVI.setup_anndata(
    cdata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal'],
)

model_C = scvi.model.SCVI(cdata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_C.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_C_dir = os.path.join(scvi_dir, 'model_C')
print(model_C_dir)
model_C.save(model_C_dir)

### model_D

scvi.model.SCVI.setup_anndata(
    cdata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal', 'doublet_scores'],
)

model_D = scvi.model.SCVI(cdata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_D.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_D_dir = os.path.join(scvi_dir, 'model_D')
print(model_D_dir)
model_D.save(model_D_dir)

### model_E

edata = cdata[cdata.obs.doublet_scores < 0.1].copy()
edata

scvi.model.SCVI.setup_anndata(
    edata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal'],
)

model_E = scvi.model.SCVI(edata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_E.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_E_dir = os.path.join(scvi_dir, 'model_E')
print(model_E_dir)
model_E.save(model_E_dir)

### model_F

fdata = adata[adata.obs.doublet_scores < 0.2].copy()
fdata

scvi.model.SCVI.setup_anndata(
    fdata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal'],
)

model_F = scvi.model.SCVI(fdata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_F.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_F_dir = os.path.join(scvi_dir, 'model_F')
print(model_F_dir)
model_F.save(model_F_dir)

### model_G

scvi.model.SCVI.setup_anndata(
    fdata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal', 'doublet_scores'],
)

model_G = scvi.model.SCVI(fdata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_G.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_G_dir = os.path.join(scvi_dir, 'model_G')
print(model_G_dir)
model_G.save(model_G_dir)

### model_H

hdata = adata[adata.obs.doublet_scores < 0.1].copy()
hdata

scvi.model.SCVI.setup_anndata(
    hdata,
    layer='counts',
    categorical_covariate_keys=['group'],
    continuous_covariate_keys=['total_counts', 'pct_counts_mt', 'pct_counts_ribosomal', 'doublet_scores'],
)

model_H = scvi.model.SCVI(hdata)

scvi.train.Trainer(accelerator='cpu', devices=1)
model_H.train()

scvi_dir = '2024-06_tbi-snseq-rivanna/scvi'
print(scvi_dir)

model_H_dir = os.path.join(scvi_dir, 'model_H')
print(model_H_dir)
model_H.save(model_H_dir)
